+++ packages/design-tokens/scripts/style-dictionary.js
@@ -1,7 +1,6 @@
 /**
  * Ported from primer/primitives, as it was removed in v8 release and we still need it
  */
-
 const glob = require('fast-glob')
 const {PrimerStyleDictionary} = require('@primer/primitives/dist/build/PrimerStyleDictionary.js')
 
@@ -48,62 +47,6 @@ StyleDictionary.registerFormat({
     return `${fileHeader({file})}
     
 module.exports = ${JSON.stringify(recursiveleyFlattenDictionary(dictionary.tokens), null, 2)}`
-  },
-})
-
-/**
- * Replacement format for typescript/module-declarations
- * Type schema corresponds to javascript/module-v2 format
- */
-StyleDictionary.registerFormat({
-  name: 'typescript/module-declarations-v2',
-  formatter({dictionary, options, file}) {
-    const {moduleName = `tokens`} = options
-
-    const getType = value => {
-      switch (typeof value) {
-        case 'string':
-          return 'string'
-        case 'number':
-          return 'number'
-        default:
-          return 'any'
-      }
-    }
-
-    const recursiveTypeGeneration = obj => {
-      const tree = {}
-      const shortHandSizes = ['large', 'medium', 'small']
-      if (typeof obj !== 'object' || Array.isArray(obj)) {
-        return obj
-      }
-
-      if (obj.hasOwnProperty('value') && typeof obj.value === 'string') {
-        return getType(obj.value)
-      } else {
-        for (const name in obj) {
-          if ((obj.hasOwnProperty(name) && obj.name === 'shorthand') || shortHandSizes.includes(obj.name)) {
-            for (const shorthandKey in obj.value) {
-              tree[shorthandKey] = getType(obj.value[shorthandKey])
-            }
-            return tree
-          } else if (obj.hasOwnProperty(name)) {
-            tree[name] = recursiveTypeGeneration(obj[name])
-          }
-        }
-      }
-      return tree
-    }
-
-    const output = `${fileHeader({file})}
-    
-declare const ${moduleName}: ${JSON.stringify(recursiveTypeGeneration(dictionary.tokens), null, 2)}
-export default ${moduleName};`
-
-    return output
-      .replace(/"any"/g, 'any')
-      .replace(/"string"/g, 'string')
-      .replace(/"number"/g, 'number')
   },
 })
 
@@ -208,7 +151,7 @@ function buildPrimitives(
       },
       js: {
         buildPath: `${outputPath}/js/`,
-        transforms: ['name/pathToPascalCase', 'pxToRem'],
+        transforms: ['name/pathToPascalCase', 'dimension/rem'],
         // map the array of token file paths to style dictionary output files
         files: files.map(filePath => {
           return {
@@ -220,7 +163,7 @@ function buildPrimitives(
       },
       jsModule: {
         buildPath: `${outputPath}/js/module/`,
-        transforms: ['pxToRem'],
+        transforms: ['dimension/rem'],
         // map the array of token file paths to style dictionary output files
         files: files.map(filePath => {
           return {
@@ -232,7 +175,7 @@ function buildPrimitives(
       },
       tsTypes: {
         buildPath: `${outputPath}/ts/`,
-        transforms: ['pxToRem'],
+        transforms: ['dimension/rem'],
         // map the array of token file paths to style dictionary output files
         files: files.map(filePath => {
           return {
@@ -244,7 +187,7 @@ function buildPrimitives(
       },
       ts: {
         buildPath: `${outputPath}/ts/`,
-        transforms: ['pxToRem'],
+        transforms: ['dimension/rem'],
         // map the array of token file paths to style dictionary output files
         files: files.map(filePath => {
           return {
